refactor(application): move mergeObjects into separate file
