+++ examples/todo-app/src/application/makeApplication.ts
@@ -1,5 +1,5 @@
 import { ApplicationSinks, ApplicationSources, Infrastructure } from './types';
-import { Stream, map, merge } from 'most';
+import { Stream, map } from 'most';
 import { filter, length } from 'ramda';
 
 import { AddTodo } from './AddTodo';
@@ -8,6 +8,7 @@ import { FilterTodos } from './FilterTodos';
 import { RemoveTodo } from './RemoveTodo';
 import { Todo } from '../domain/model';
 import { UpdateTodo } from './UpdateTodo';
+import { mergeObjects } from './mergeObjects';
 import { proxy } from 'most-proxy';
 
 export function makeApplication(infrastructure: Infrastructure) {
@@ -44,28 +45,6 @@ export function makeApplication(infrastructure: Infrastructure) {
       ...itemCounts(todos$),
     };
   };
-}
-
-function mergeObjects<A extends { [key: string]: Stream<any> }>(...objs: Array<A>): A {
-  if (objs.length === 1)
-    return objs[0];
-
-  const mergedObject = {} as A;
-
-  objs.forEach((obj: A) => {
-    const keys = Object.keys(obj);
-
-    keys.forEach((key: keyof A) => {
-      const value = obj[key];
-
-      if (!mergedObject[key])
-        mergedObject[key] = value;
-      else
-        mergedObject[key] = merge(mergedObject[key], value);
-    });
-  });
-
-  return mergedObject;
 }
 
 function itemCounts(todos$: Stream<Array<Todo>>) {
